afanasy: all states and flags integers are 64 bit (more bits needed) task has ERROR_READY state: when it has an error but it was restarted, and now its ready (but not started yet, or can't start)

+++ afanasy/src/libafanasy/taskprogress.h
@@ -19,7 +19,7 @@ public:
 
    virtual int calcWeight() const;
 
-   uint32_t state;         ///< state per block per task.
+   int64_t  state;         ///< state per block per task.
    int8_t   percent;       ///< percent per block per task.
    int64_t  frame;         ///< frame per block per task.
    int8_t   percentframe;  ///< frame percent per block per task.
